extract fixtures and snapshot plugin from tests
