+++ localstack/utils/testing/snapshots/snapshots.py
