+++ tests/integration/test_s3.py
@@ -52,7 +52,7 @@ from localstack.utils.common import (
     to_str,
 )
 from localstack.utils.server import http2_server
-from tests.integration.fixtures import only_localstack
+from localstack.utils.testing.utils import only_localstack
 
 TEST_BUCKET_NAME_WITH_POLICY = "test-bucket-policy-1"
 TEST_BUCKET_WITH_VERSIONING = "test-bucket-versioning-1"
